refactor(frontend): consolidate form and ui components
