+++ frontend/src/components/form/index.ts
@@ -1,7 +1,4 @@
 export { BatchFormControls } from "./BatchFormControls";
-export { ColorContrastMessage } from "./ColorContrastMessage";
-export { ColorControls } from "./ColorControls";
 export { CoverForm } from "./CoverForm";
 export { CoverFormControls } from "./CoverFormControls";
 export { CoverPreviewDisplay } from "./CoverPreviewDisplay";
-export { FormField } from "./FormField";
