--- frontend/src/components/ui/ColorPicker.tsx
@@ -1,19 +1,0 @@
-import { cn } from "@/lib";
-
-interface ColorPickerProps
-	extends React.InputHTMLAttributes<HTMLInputElement> {}
-
-export function ColorPicker({ className, title, ...props }: ColorPickerProps) {
-	return (
-		<input
-			type="color"
-			className={cn(
-				"h-10 w-full rounded-md border border-input p-0 cursor-pointer",
-				"focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2",
-				className,
-			)}
-			title={title || "Select a color"}
-			{...props}
-		/>
-	);
-}
