--- frontend/src/components/ui/Input.tsx
@@ -1,18 +1,0 @@
-import { cn } from "@/lib";
-
-interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {}
-
-export function Input({ className, type = "text", ...props }: InputProps) {
-	return (
-		<input
-			type={type}
-			className={cn(
-				"flex h-10 w-full rounded-xl border border-gray-300 bg-gray-50 px-3 py-2 text-sm",
-				"placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2",
-				"disabled:cursor-not-allowed disabled:opacity-50",
-				className,
-			)}
-			{...props}
-		/>
-	);
-}
