--- frontend/src/components/ui/Select.tsx
@@ -1,19 +1,0 @@
-import { cn } from "@/lib";
-
-interface SelectProps extends React.SelectHTMLAttributes<HTMLSelectElement> {}
-
-export function Select({ className, children, ...props }: SelectProps) {
-	return (
-		<select
-			className={cn(
-				"flex h-10 w-full rounded-xl border border-gray-300 bg-gray-50 px-3 py-2 text-sm",
-				"focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2",
-				"disabled:cursor-not-allowed disabled:opacity-50",
-				className,
-			)}
-			{...props}
-		>
-			{children}
-		</select>
-	);
-}
