+++ frontend/src/components/ui/index.ts
@@ -1,9 +1,5 @@
 export { Button } from "./Button";
 export { Card, KPICard } from "./Cards";
-export { ColorPicker } from "./ColorPicker";
-export { FormError } from "./FormError";
-export { Input } from "./Input";
-export { Nav } from "./Nav";
+export { ColorPicker, FormError, Input, Select } from "./FormControls";
 export { SectionTitle } from "./SectionTitle";
-export { Select } from "./Select";
 export { Skeleton } from "./Skeleton";
