+++ package.json
@@ -22,6 +22,10 @@
 		"lint:api": "biome check api/",
 		"lint:frontend": "biome check frontend/",
 		"lint:shared": "biome check shared/",
+		"lint:fix": "biome check --write frontend/ api/ shared/",
+		"lint:fix:api": "biome check --write api/",
+		"lint:fix:frontend": "biome check --write frontend/",
+		"lint:fix:shared": "biome check --write shared/",
 		"lint:md": "npx markdownlint-cli \"**/*.md\" --ignore \"**/node_modules/**\"",
 		"deps:deprecated": "npm ls --all --json --long 2>/dev/null | rg '\"deprecated\"' -n -C 4",
 		"format": "biome format --write frontend/ api/ shared/",
